+++ src/onegov.org/onegov/org/theme/org_theme.py
@@ -16,7 +16,7 @@ class OrgTheme(BaseTheme):
 
     # don't touch this number, it's incremented using bumpversion, so every
     # release will automatically trigger a rebuild of the theme
-    version = '0.40.8'
+    version = '0.41.0'
 
     @property
     def default_options(self):